### 2019-10-04 #1783 @whedon
[PRE REVIEW]: pylustrator: code generation for reproducible figures for publication
**Submitting author:** @rgerum (<a href="http://orcid.org/0000-0001-5893-2650">Richard Gerum</a>)
**Repository:** <a href="https://github.com/rgerum/pylustrator" target ="_blank">https://github.com/rgerum/pylustrator</a>
**Version:** v0.12.0
