### 2019-10-07 de42118
load: when loading a script from a different folder, execute the script with its containing folder as working directory
