+++ pylustrator/helper_functions.py
@@ -27,7 +27,7 @@ import traceback
 from .parse_svg import svgread
 from matplotlib.axes._subplots import Axes
 from .pyjack import replace_all_refs
-
+import os
 
 def fig_text(x, y, text, unit="cm", *args, **kwargs):
     """
@@ -131,134 +131,164 @@ def imShowFullFigure(im, filename, fig1, dpi):
     for spine in ["left", "right", "top", "bottom"]:
         ax.spines[spine].set_visible(False)
 
+
+class changeFolder:
+    """
+    An environment that changes the working directory
+    """
+    def __init__(self, directory):
+        self.directory = directory
+
+    def __enter__(self):
+        self.old_dir = os.getcwd()
+        if self.directory:
+            os.chdir(self.directory)
+
+    def __exit__(self, type, value, traceback):
+        os.chdir(self.old_dir)
+
 def loadFigureFromFile(filename, fig1=None, offset=None, dpi=None, cache=True):
     from matplotlib import rcParams
     from pylustrator import changeFigureSize
     import pylustrator
-    import os
 
-    # defaults to the current figure
-    if fig1 is None:
-        fig1 = plt.gcf()
+    # change to the directory of the filename (to execute the code relative to this directory)
+    dirname, filename = os.path.split(filename)
+    with changeFolder(dirname):
+        if dirname:
+            os.chdir(dirname)
 
-    class noShow:
-        """
-        An environment that prevents the script from calling the plt.show function
-        """
-        def __enter__(self):
-            # store the show function
-            self.show = plt.show
-            self.dragger = pylustrator.start
+        # defaults to the current figure
+        if fig1 is None:
+            fig1 = plt.gcf()
 
-            # define an empty function
-            def empty(*args, **kwargs):
-                pass
+        class noShow:
+            """
+            An environment that prevents the script from calling the plt.show function
+            """
+            def __enter__(self):
+                # store the show function
+                self.show = plt.show
+                self.dragger = pylustrator.start
 
-            # set the show function to the empty function
-            plt.show = empty
-            pylustrator.start = empty
+                # define an empty function
+                def empty(*args, **kwargs):
+                    pass
 
-        def __exit__(self, type, value, traceback):
-            # restore the old show function
-            plt.show = self.show
-            pylustrator.start = self.dragger
+                # set the show function to the empty function
+                plt.show = empty
+                pylustrator.start = empty
 
-    class noNewFigures:
-        """
-        An environment that prevents the script from creating new figures in the figure manager
-        """
-        def __enter__(self):
-            fig = plt.gcf()
-            self.fig = plt.figure
-            figsize = rcParams['figure.figsize']
-            fig.set_size_inches(figsize[0], figsize[1])
-            def figure(num=None, figsize=None, *args, **kwargs):
+            def __exit__(self, type, value, traceback):
+                # restore the old show function
+                plt.show = self.show
+                pylustrator.start = self.dragger
+
+        class noNewFigures:
+            """
+            An environment that prevents the script from creating new figures in the figure manager
+            """
+            def __enter__(self):
                 fig = plt.gcf()
-                if figsize is not None:
-                    fig.set_size_inches(figsize[0], figsize[1], forward=True)
-                return fig
-            plt.figure = figure
+                self.fig = plt.figure
+                figsize = rcParams['figure.figsize']
+                fig.set_size_inches(figsize[0], figsize[1])
+                def figure(num=None, figsize=None, *args, **kwargs):
+                    fig = plt.gcf()
+                    if figsize is not None:
[…]
