### 2019-10-07 94281f0
load: fixed loading figures with texts
