### 2019-10-08 dfab9cd
fixed saving tick labels (fixes #3)
