### 2019-10-10 51bbb8e
added alignment buttons
