### 2019-10-10 4f3b061
despine button also works for group selections
