### 2019-10-10 2b3abaa
update axes selection rectangles on plot resize (fixes #8)
