### 2019-10-11 d562612
added arxiv citation to docs
