### 2019-10-22 #1783 @labarba
@rgerum — This is a re-submission of the same software that was considered out-of-scope on June 21:
https://github.com/openjournals/joss-reviews/issues/1499#issuecomment-504573661

Can you explain here what is different in the current submission, if anything, to the previous one?
