### 2019-10-22 #1783 @rgerum
I rewrote the paper completely to stress more that the software is a tool in its own right and not just a small extension of the matplotlib package. Pylustrator is not a tool that extends the possible plotting functions of matplotlib, but is a tool to generate python code that composes single plots into complex composite figures for publication.
As I observed that most scientists I know compose their figures after plotting in image manipulation software (e.g. GIMP, Photoshop, etc.) which is error prone (accidentally altering the data points) and not reproducible (for other and also for oneself if the evaluation needs to be adapted later on). Out of this need I developed this software that aims to provide a reproducible tool to achieve the figure composition with code.
As the software is already used in a lot of publications, I want to have a citable paper and a way to further promote it, as I think it is a valuable contribution for science.

As I see other papers in JOSS that are also focusing on visualizations, e.g.
https://joss.theoj.org/papers/10.21105/joss.01425
https://joss.theoj.org/papers/10.21105/joss.01628
I do not see why my submission could not fit into JOSS.
