### 2019-10-27 #1783 @labarba
👋 @tacaswell — Would you be able to review this submission for JOSS? Seems right up your alley...
**pylustrator: code generation for reproducible figures for publication**
https://github.com/rgerum/pylustrator
