### 2019-10-27 #1783 @labarba
👋 @aokad — Are you interested in reviewing this JOSS submission? You signed up with "visualization" as your area of interest…
**pylustrator: code generation for reproducible figures for publication**
https://github.com/rgerum/pylustrator
