### 2019-10-31 af1b9a6
load: fixed loading plots from different directories
