### 2019-10-31 1059ee4
change_tracker: added collections to referenceable items
