### 2019-11-27 #1783 @labarba
👋  @aokad — You have listed visualization as your topic for reviewing... would you be willing to review this submission for JOSS?
**pylustrator: code generation for reproducible figures for publication**
https://github.com/rgerum/pylustrator
