### 2019-11-27 #1783 @labarba
👋  @alexanderfurnas — I notice you listed visualization as one of your review topics... would you be interested in contributing a JOSS review for this submission?
**pylustrator: code generation for reproducible figures for publication**
https://github.com/rgerum/pylustrator

(BTW, I found that you entered your handle as `zfurnas`, your Twitter handle rather than GitHub, which is probably why you haven't been asked to review all this time!)
