### 2019-11-28 #1783 @labarba
To potential reviewers that said you are unable to review: please unsubscribe to this thread to save your inbox!
@aokad : I just noticed that I pinged you twice, after you noted your unavailability... Oops, disregard!

@tacaswell, @alexanderfurnas — let us know if you might be able to contribute a review for this JOSS submission. Thanks!
