### 2019-12-20 ff5d6df
Axes labels always use get_xaxis().get_label().set_text instead of set_xlabel (fixes #9)
