### 2020-01-01 #1783 @labarba
👋  @rgerum — Apologies for the slow-moving review here. I had a hard time finding willing reviewers, and then dropped the ball, I'm afraid. We now have one reviewer, and am looking for a second. Stay tuned!

Thank you for agreeing to review, @story645! Have a look at the [JOSS Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let me know if you have any questions. We'll need to add a second reviewer before starting the Review issue, but in the meantime you can look through the software and start your assessment.
