### 2020-01-09 a70aca2
added legend styling options (fixes #6)
