+++ pylustrator/QtGuiDrag.py
@@ -560,7 +560,8 @@ class ComboWidget(QtWidgets.QWidget, Linkable):
         return "\""+str(self.get())+"\""
 
 
-class CheckWidget(QtWidgets.QWidget):
+class CheckWidget(QtWidgets.QWidget, Linkable):
+    editingFinished = QtCore.Signal()
     stateChanged = QtCore.Signal(int)
     noSignal = False
 
@@ -580,6 +581,7 @@ class CheckWidget(QtWidgets.QWidget):
     def onStateChanged(self):
         if not self.noSignal:
             self.stateChanged.emit(self.input1.isChecked())
+            self.editingFinished.emit()
 
     def setChecked(self, state):
         self.noSignal = True
@@ -588,6 +590,15 @@ class CheckWidget(QtWidgets.QWidget):
 
     def isChecked(self):
         return self.input1.isChecked()
+
+    def get(self):
+        return self.isChecked()
+
+    def set(self, value):
+        self.setChecked(value)
+
+    def getSerialized(self):
+        return "True" if self.get() else "False"
 
 
 class RadioWidget(QtWidgets.QWidget):
@@ -903,6 +914,93 @@ class TextPropertiesWidget(QtWidgets.QWidget):
                 element.set_fontsize(value)
                 element.figure.change_tracker.addChange(element, ".set_fontsize(%d)" % value)
             self.target.figure.canvas.draw()
+
+
+class LegendPropertiesWidget(QtWidgets.QWidget):
+    stateChanged = QtCore.Signal(int, str)
+    noSignal = False
+    target_list = None
+
+    def __init__(self, layout):
+        QtWidgets.QWidget.__init__(self)
+        layout.addWidget(self)
+        self.layout = QtWidgets.QVBoxLayout(self)
+        self.layout.setContentsMargins(0, 0, 0, 0)
+
+        self.property_names = [
+            ("frameon", "frameon", bool),
+            ("borderpad", "borderpad", float),
+            ("labelspacing", "labelspacing", float),
+            ("handlelength", "handlelength", float),
+            ("handletextpad", "handletextpad", float),
+            ("columnspacing", "columnspacing", float),
+            ("markerscale", "markerscale", float),
+            ("ncol", "_ncol", int),
+        ]
+        self.properties = {}
+
+        self.widgets = {}
+        for index, (name, name2, type_) in enumerate(self.property_names):
+            if index % 3 == 0:
+                layout = QtWidgets.QHBoxLayout()
+                layout.setContentsMargins(0, 0, 0, 0)
+                self.layout.addLayout(layout)
+            if type_ == bool:
+                widget = CheckWidget(layout, name+":")
+                widget.editingFinished.connect(lambda name=name, widget=widget: self.changePropertiy(name, widget.get()))
+            else:
+                label = QtWidgets.QLabel(name+":")
+                layout.addWidget(label)
+                if type_ == float:
+                    widget = QtWidgets.QDoubleSpinBox()
+                    widget.setSingleStep(0.1)
+                elif type_ == int:
+                    widget = QtWidgets.QSpinBox()
+                layout.addWidget(widget)
+                widget.valueChanged.connect(lambda x, name=name: self.changePropertiy(name, x))
+            self.widgets[name] = widget
+
+    def changePropertiy(self, name, value):
+        if self.target is None:
+            return
+
+        bbox = self.target.get_frame().get_bbox()
+        self.properties[name] = value
+        axes = self.target.axes
+        axes.legend(**self.properties)
+        self.target = axes.get_legend()
+        fig = self.target.figure
+        fig.change_tracker.addChange(axes, ".legend(%s)" % (", ".join("%s=%s" % (k, v) for k, v in self.properties.items())))
[…]
