### 2020-01-20 37c1500
set_xscale is sorted before set_xlim (fixes #12)
