### 2020-01-21 91ef682
Legend editing, added property to add a legend title
