### 2020-01-26 #1989 @labarba
Hi folks! @story645, @tacaswell — How is your review going? I don't see items ticked off on your checklists. Do you have any questions for us?
