### 2020-02-03 63128ca
legend editing, fixed bug when no title was specified
