### 2020-02-03 2741678
legend editing, added fontsize property
