### 2020-02-03 86068dd
legend editing only saves parameter values that deviate from the defaults
