### 2020-02-03 d38a43c
added short overview directly to the readme.rst
