### 2020-02-03 2e4b376
added title_fontsize property for legend editing (fixes #13)
