### 2020-02-03 58637a7
allow to despine multiple axes simultaneously (fixes #14)
