### 2020-02-04 rgerum/pylustrator#17 @tacaswell
Missing community guidelines
As part of the JOSS review I could not find the community guide lines for contributing.  The docs do say how to report bugs (although the text refers to bitbucket but links to github!).
