### 2020-02-04 #1989 @tacaswell
First, this is a super cool project, I am very impressed with the functionality.

 - Created https://github.com/rgerum/pylustrator/issues/17 to track missing community guide lines
 - This software is primarily a GUI interface to editing Matplotlib figures (which is pretty awesome!) and the intended usage is through the UI, not as a library.  There is not documentation on the API (because there is not one) nor automated tests.   Poking through the code there is a lack of docstrings on a majority of the functions.  While I don't think that it is worth doing a full GUI mock test, I think it would be reasonable to test some of the functionality (such as ensuring the the monkey patching works as expected, that the file patching works as expected for a manually generated change set etc).
 - This sort of user-action-recording functionality exist in several other visualization / image analysis packages (paraview/vtk/and friends, avizo, dragonfly, etc).  I know of no other tool that provides this functionality on top of Matplotlib, however I think that the novelty claim in the conclusions of the paper are too strong.
