### 2020-02-04 d1949e4
Replaced reference to Bitbucket with Github in the documentation.
