+++ readme.rst
@@ -1,28 +1,0 @@
-.. -*- mode: rst -*-
-
-Pylustrator
-===========
-
-|DOC|_ |License|_ |DOI|_
-
-.. |DOC| image:: https://readthedocs.org/projects/pylustrator/badge/
-.. _DOC: https://pylustrator.readthedocs.io
-
-.. |License| image:: https://img.shields.io/badge/License-GPLv3-blue.svg
-.. _License: http://www.gnu.org/licenses/gpl-3.0.html
-
-.. |DOI| image:: https://img.shields.io/badge/DOI-10.5281/zenodo.1294663-blue.svg
-.. _DOI: https://zenodo.org/record/1294664
-
-.. figure:: docs/images/logo.png
-    :align: left
-
-Pylustrator is a software to prepare your figures for publication in a reproducible way. This means you receive a figure
-representing your data and alongside a generated code file that can exactly reproduce the figure as you put them in the
-publication, without the need to readjust things in external programs.
-
-Pylustrator offers an interactive interface to find the best way to present your data in a figure for publication.
-Added formatting an styling can be saved by automatically generated code. To compose multiple figures to panels,
-pylustrator can compose different subfigures to a single figure.
-
-Please also refer to the `Documentation <https://pylustrator.readthedocs.io>`_ for more information.
