+++ README.md
@@ -1,0 +1,35 @@
+# Pylustrator
+
+[![DOC](image:: https://readthedocs.org/projects/pylustrator/badge/)](https://pylustrator.readthedocs.io)
+[![License](image:: https://img.shields.io/badge/License-GPLv3-blue.svg)](http://www.gnu.org/licenses/gpl-3.0.html)
+[![DOI](https://img.shields.io/badge/DOI-10.5281/zenodo.1294663-blue.svg)](https://zenodo.org/record/1294664)
+
+
+<img style="float: left;" alt="docs/images/logo.png" src="docs/images/logo.png" />
+
+Pylustrator is a software to prepare your figures for publication in a reproducible way. This means you receive a figure
+representing your data and alongside a generated code file that can exactly reproduce the figure as you put them in the
+publication, without the need to readjust things in external programs.
+
+Pylustrator offers an interactive interface to find the best way to present your data in a figure for publication.
+Added formatting an styling can be saved by automatically generated code. To compose multiple figures to panels,
+pylustrator can compose different subfigures to a single figure.
+
+Please also refer to the [Documentation](https://pylustrator.readthedocs.io) for more information.
+
+## Issues, Questions, and Suggestions
+
+Please submit your questions, suggestions, and bug reports to the
+[Issue Tracker](https://github.com/rgerum/pylustrator/issues)
+
+
+## Contributing
+
+You want to contribute? Great!
+Contributing works best if you creat a pull request with your changes.
+
+1. Fork it!
+2. Create your feature branch: `git checkout -b my-new-feature`
+3. Commit your changes: `git commit -am 'Add some feature'`
+4. Push to the branch: `git push origin my-new-feature`
+5. Submit a pull request!
