### 2020-02-04 rgerum/pylustrator#17 @rgerum
I have added a short contribution statement to the README.md and changed the text in the documentation to properly refer to Github.
