### 2020-02-04 9811ef2
added test to test the code generation for moving an axes
