### 2020-02-04 dc7b81a
try to prevent crashes in tests
