### 2020-02-04 595894e
allow to despine multiple axes simultaneously (fixes #14)
