### 2020-02-05 31ff4ff
removed signals with "PyQt_PyObject"
