### 2020-02-05 a82f36b
on test cleanup don't throw error if .tmp file does not exist
