### 2020-02-05 b307c63
snap.py: added docstrings
