+++ pylustrator/snap.py
@@ -19,14 +19,17 @@
 # You should have received a copy of the GNU General Public License
 # along with Pylustrator. If not, see <http://www.gnu.org/licenses/>
 
+from typing import List, Tuple, Optional
+
+import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
-from matplotlib.lines import Line2D
+from matplotlib.artist import Artist
 from matplotlib.axes._subplots import Axes
+from matplotlib.legend import Legend
+from matplotlib.lines import Line2D
 from matplotlib.patches import Rectangle, Ellipse, FancyArrowPatch
 from matplotlib.text import Text
-from matplotlib.legend import Legend
-import matplotlib as mpl
 
 DIR_X0 = 1
 DIR_Y0 = 2
@@ -34,41 +37,39 @@ DIR_X1 = 4
 DIR_Y1 = 8
 
 
-def get_loc_in_canvas(legend):
-    offsetbox = legend._legend_box
-    renderer = offsetbox.figure._cachedRenderer
-    w, h, xd, yd = offsetbox.get_extent(renderer)
-    ox, oy = offsetbox._offset()
-    loc_in_canvas = (ox - xd, oy - yd)
-
-    return loc_in_canvas
-
-
-def checkXLabel(target):
+def checkXLabel(target: Artist):
+    """ checks if the target is the xlabel of an axis """
     for axes in target.figure.axes:
         if axes.xaxis.get_label() == target:
             return axes
 
-def checkYLabel(target):
+
+def checkYLabel(target: Artist):
+    """ checks if the target is the ylabel of an axis """
     for axes in target.figure.axes:
         if axes.yaxis.get_label() == target:
             return axes
 
 
 class TargetWrapper(object):
+    """ a wrapper to add unified set and get position methods for any matplotlib artist """
     target = None
 
-    def __init__(self, target):
+    def __init__(self, target: Artist):
         self.target = target
         self.figure = target.figure
         self.do_scale = True
         self.fixed_aspect = False
+        # a patch uses the data_transform
         if isinstance(self.target, mpl.patches.Patch):
             self.get_transform = self.target.get_data_transform
+        # axes use the figure_transform
         elif isinstance(self.target, Axes):
+            # and optionally have a fixed aspect ratio
             if self.target.get_aspect() != "auto" and self.target.get_adjustable() != "datalim":
                 self.fixed_aspect = True
             self.get_transform = lambda: self.target.figure.transFigure
+        # texts use get_transform
         elif isinstance(self.target, Text):
             if getattr(self.target, "xy", None) is not None:
                 self.do_scale = True
@@ -77,19 +78,23 @@ class TargetWrapper(object):
             if checkXLabel(self.target):
                 self.label_factor = self.figure.dpi / 72.0
                 if getattr(self.target, "pad_offset", None) is None:
-                    self.target.pad_offset = self.target.get_position()[1] - checkXLabel(self.target).xaxis.labelpad * self.label_factor
+                    self.target.pad_offset = self.target.get_position()[1] - checkXLabel(
+                        self.target).xaxis.labelpad * self.label_factor
                 self.label_y = self.target.get_position()[1]
             elif checkYLabel(self.target):
                 self.label_factor = self.figure.dpi / 72.0
                 if getattr(self.target, "pad_offset", None) is None:
-                    self.target.pad_offset = self.target.get_position()[0] - checkYLabel(self.target).yaxis.labelpad * self.label_factor
+                    self.target.pad_offset = self.target.get_position()[0] - checkYLabel(
+                        self.target).yaxis.labelpad * self.label_factor
                 self.label_x = self.target.get_position()[0]
             self.get_transform = self.target.get_transform
+        # the default is to use get_transform
         else:
[…]
