+++ pylustrator/drag_bib.py
@@ -1,1393 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# drag_bib.py
-
-# Copyright (c) 2016-2019, Richard Gerum
-#
-# This file is part of Pylustrator.
-#
-# Pylustrator is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Pylustrator is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Pylustrator. If not, see <http://www.gnu.org/licenses/>
-
-from __future__ import division, print_function
-import numpy as np
-import traceback
-import matplotlib.pyplot as plt
-from matplotlib.text import Text
-from matplotlib.lines import Line2D
-from matplotlib.patches import Rectangle, Ellipse
-from matplotlib.figure import Figure
-from matplotlib.axes._subplots import Axes
-import matplotlib
-import uuid
-import re
-from natsort import natsorted
-
-DIR_X0 = 1
-DIR_Y0 = 2
-DIR_X1 = 4
-DIR_Y1 = 8
-
-def getReference(element):
-    if element is None:
-        return ""
-    if isinstance(element, Figure):
-        return "fig"
-    if isinstance(element, matplotlib.lines.Line2D):
-        index = element.axes.lines.index(element)
-        return getReference(element.axes)+".lines[%d]" % index
-    if isinstance(element, matplotlib.text.Text):
-        if element.axes:
-            index = element.axes.texts.index(element)
-            return getReference(element.axes)+".texts[%d]" % index
-        index = element.figure.texts.index(element)
-        return "fig.texts[%d]" % (index)
-    if isinstance(element, matplotlib.patches.Patch):
-        if element.axes:
-            index = element.axes.patches.index(element)
-            return getReference(element.axes)+".patches[%d]" % index
-        index = element.figure.patches.index(element)
-        return "fig.patches[%d]" % (index)
-    if isinstance(element, matplotlib.axes._axes.Axes):
-        if element.get_label():
-            return "fig.ax_dict[\"%s\"]" % element.get_label()
-        return "fig.axes[%d]" % element.number
-    if isinstance(element, matplotlib.legend.Legend):
-        return getReference(element.axes)+".get_legend()"
-    raise(TypeError(str(type(element))+" not found"))
-
-
-class FigureDragger:
-    last_picked = False
-    active_object = None
-    drag_object = None
-    fig = None
-    first_resize = True
-    displaying = False
-    snaps = None
-
-    changes = None
-    saved = True
-
-    def __init__(self, fig, xsnaps=None, ysnaps=None, unit="cm"):
-        self.fig = fig
-        # store dragger, so that it is not eaten by the garbage collector
-        fig.figure_dragger = self
-        self.grabbers = []
-        self.edits = []
-        self.last_edit = -1
-        self.changes = {}
-
-        # make all the subplots pickable
-        for index, axes in enumerate(self.fig.axes):
-            #axes.set_title(index)
-            axes.number = index
-            axes.set_picker(True)
-            leg = axes.get_legend()
-            if leg:
-                self.make_dragable(leg)
-            for text in axes.texts:
-                self.make_dragable(text)
-
-            self.make_dragable(axes)
-        for text in self.fig.texts:
-            self.make_dragable(text)
-
[…]
