+++ pylustrator/__init__.py
@@ -19,7 +19,6 @@
 # You should have received a copy of the GNU General Public License
 # along with Pylustrator. If not, see <http://www.gnu.org/licenses/>
 
-from .drag_bib import StartPylustrator, moveArtist
 from .QtGuiDrag import initialize as start
 from .helper_functions import fig_text, add_axes, add_image, despine, changeFigureSize, mark_inset, VoronoiPlot, selectRectangle, mark_inset_pos, draw_from_point_to_bbox, draw_from_point_to_point, loadFigureFromFile, add_letter, add_letters
 from .QtGui import initialize as StartColorChooser
