### 2020-02-05 17e1690
added doc strings to ax_rasterisation.py
