### 2020-02-05 f2ffe98
change_tracker.py added docstrings
