+++ pylustrator/change_tracker.py
@@ -19,29 +19,29 @@
 # You should have received a copy of the GNU General Public License
 # along with Pylustrator. If not, see <http://www.gnu.org/licenses/>
 
-from __future__ import division, print_function
-
 import re
 import sys
 import traceback
+from typing import IO
 
 import matplotlib
 import matplotlib as mpl
-import matplotlib.pyplot as plt
 from matplotlib import _pylab_helpers
+from matplotlib.artist import Artist
 from matplotlib.axes._subplots import Axes
+from matplotlib.collections import Collection
 from matplotlib.figure import Figure
 from matplotlib.lines import Line2D
 from matplotlib.patches import Rectangle
 from matplotlib.text import Text
-from matplotlib.collections import Collection
 from natsort import natsorted
 
 from .exception_swallower import Dummy
 from .jupyter_cells import open
 
 
-def getReference(element, allow_using_variable_names=True):
+def getReference(element: Artist, allow_using_variable_names=True):
+    """ get the code string that represents the given Artist. """
     if element is None:
         return ""
     if isinstance(element, Figure):
@@ -121,7 +121,8 @@ def getReference(element, allow_using_variable_names=True):
     raise TypeError(str(type(element)) + " not found")
 
 
-def setFigureVariableNames(figure):
+def setFigureVariableNames(figure: Figure):
+    """ get the global variable names that refer to the given figure """
     import inspect
     mpl_figure = _pylab_helpers.Gcf.figs[figure].canvas.figure
     calling_globals = inspect.stack()[2][0].f_globals
@@ -137,10 +138,11 @@ def setFigureVariableNames(figure):
 
 
 class ChangeTracker:
+    """ a class that records a list of the change to the figure """
     changes = None
     saved = True
 
-    def __init__(self, figure):
+    def __init__(self, figure: Figure):
         global stack_position
         self.figure = figure
         self.edits = []
@@ -159,7 +161,8 @@ class ChangeTracker:
 
         self.load()
 
-    def addChange(self, command_obj, command, reference_obj=None, reference_command=None):
+    def addChange(self, command_obj: Artist, command: str, reference_obj: Artist = None, reference_command: str = None):
+        """ add a change """
         command = command.replace("\n", "\\n")
         if reference_obj is None:
             reference_obj = command_obj
@@ -168,7 +171,8 @@ class ChangeTracker:
         self.changes[reference_obj, reference_command] = (command_obj, command)
         self.saved = False
 
-    def removeElement(self, element):
+    def removeElement(self, element: Artist):
+        """ remove an Artis from the figure """
         # create_key = key+".new"
         created_by_pylustrator = (element, ".new") in self.changes
         # delete changes related to this element
@@ -183,13 +187,15 @@ class ChangeTracker:
             element.remove()
         self.figure.selection.remove_target(element)
 
-    def addEdit(self, edit):
+    def addEdit(self, edit: list):
+        """ add an edit to the stored list of edits """
         if self.last_edit < len(self.edits) - 1:
             self.edits = self.edits[:self.last_edit + 1]
         self.edits.append(edit)
         self.last_edit = len(self.edits) - 1
 
     def backEdit(self):
+        """ undo an edit in the list """
         if self.last_edit < 0:
             return
         edit = self.edits[self.last_edit]
@@ -198,6 +204,7 @@ class ChangeTracker:
         self.figure.canvas.draw()
 
     def forwardEdit(self):
+        """ redo an edit """
         if self.last_edit >= len(self.edits) - 1:
             return
         edit = self.edits[self.last_edit + 1]
@@ -206,6 +213,7 @@ class ChangeTracker:
[…]
