### 2020-02-06 e5bbf88
drag_helper.py added docstrings
