+++ pylustrator/drag_helper.py
@@ -22,10 +22,14 @@
 from __future__ import division, print_function
 import numpy as np
 import matplotlib.pyplot as plt
+from matplotlib.artist import Artist
+from matplotlib.figure import Figure
 from matplotlib.text import Text
 from matplotlib.patches import Rectangle, Ellipse
+from matplotlib.backend_bases import MouseEvent, KeyEvent
+from typing import Sequence
 
-from .snap import TargetWrapper, getSnaps, checkSnaps, checkSnapsActive
+from .snap import TargetWrapper, getSnaps, checkSnaps, checkSnapsActive, SnapBase
 from .change_tracker import ChangeTracker
 
 DIR_X0 = 1
@@ -35,6 +39,7 @@ DIR_Y1 = 8
 
 
 class GrabFunctions(object):
+    """ basic functionality used by all grabbers """
     figure = None
     target = None
     dir = None
@@ -42,32 +47,36 @@ class GrabFunctions(object):
 
     got_artist = False
 
-    def __init__(self, parent, dir, no_height=False):
+    def __init__(self, parent, dir: int, no_height=False):
         self.figure = parent.figure
         self.parent = parent
         self.dir = dir
         self.snaps = []
         self.no_height = no_height
 
-    def on_motion(self, evt):
+    def on_motion(self, evt: MouseEvent):
+        """ callback when the object is moved """
         if self.got_artist:
             self.movedEvent(evt)
             self.moved = True
 
-    def button_press_event(self, evt):
+    def button_press_event(self, evt: MouseEvent):
+        """ when the mouse is pressed """
         self.got_artist = True
         self.moved = False
 
         self._c1 = self.figure.canvas.mpl_connect('motion_notify_event', self.on_motion)
         self.clickedEvent(evt)
 
-    def button_release_event(self, event):
+    def button_release_event(self, event: MouseEvent):
+        """ when the mouse is released """
         if self.got_artist:
             self.got_artist = False
             self.figure.canvas.mpl_disconnect(self._c1)
             self.releasedEvent(event)
 
-    def clickedEvent(self, event):
+    def clickedEvent(self, event: MouseEvent):
+        """ when the mouse is clicked """
         self.parent.start_move()
         self.mouse_xy = (event.x, event.y)
 
@@ -77,14 +86,16 @@ class GrabFunctions(object):
 
         self.snaps = getSnaps(self.targets, self.dir, no_height=self.no_height)
 
-    def releasedEvent(self, event):
+    def releasedEvent(self, event: MouseEvent):
+        """ when the mouse is released """
         for snap in self.snaps:
             snap.remove()
         self.snaps = []
 
         self.parent.end_move()
 
-    def movedEvent(self, event):
+    def movedEvent(self, event: MouseEvent):
+        """ when the mouse is moved """
         if len(self.targets) == 0:
             return
 
@@ -100,11 +111,11 @@ class GrabFunctions(object):
 class GrabbableRectangleSelection(GrabFunctions):
     grabbers = None
 
-    def addGrabber(self, x, y, dir, GrabberClass):
+    def addGrabber(self, x: float, y: float, dir: int, GrabberClass: object):
         # add a grabber object at the given coordinates
         self.grabbers.append(GrabberClass(self, x, y, dir))
 
-    def __init__(self, figure):
+    def __init__(self, figure: Figure):
         self.grabbers = []
         pos = [0, 0, 0, 0]
         self.positions = np.array(pos, dtype=float)
@@ -130,7 +141,7 @@ class GrabbableRectangleSelection(GrabFunctions):
 
         self.hide_grabber()
 
-    def add_target(self, target):
+    def add_target(self, target: Artist):
[…]
