### 2020-02-07 1eff641
added docstrings to exception_swallower.py and helper_functions.py
