+++ pylustrator/helper_functions.py
@@ -26,10 +26,13 @@ import numpy as np
 import traceback
 from .parse_svg import svgread
 from matplotlib.axes._subplots import Axes
+from matplotlib.figure import Figure
 from .pyjack import replace_all_refs
 import os
+from typing import Sequence, Union
 
-def fig_text(x, y, text, unit="cm", *args, **kwargs):
+
+def fig_text(x: float, y: float, text: str, unit: str = "cm", *args, **kwargs):
     """
     add a text to the figure positioned in cm
     """
@@ -44,7 +47,7 @@ def fig_text(x, y, text, unit="cm", *args, **kwargs):
     return fig.text(x, y, text, picker=True, *args, **kwargs)
 
 
-def add_axes(dim, unit="cm", *args, **kwargs):
+def add_axes(dim: Sequence, unit: str = "cm", *args, **kwargs):
     """
     add an axes with dimensions specified in cm
     """
@@ -62,13 +65,17 @@ def add_axes(dim, unit="cm", *args, **kwargs):
     return plt.axes([x, y, w, h], *args, **kwargs)
 
 
-def add_image(filename):
+def add_image(filename: str):
+    """ add an image to the current axes """
     plt.imshow(plt.imread(filename))
     plt.xticks([])
     plt.yticks([])
 
 
-def changeFigureSize(w, h, cut_from_top=False, cut_from_left=False, fig=None):
+def changeFigureSize(w: float, h: float, cut_from_top: bool = False, cut_from_left: bool = False, fig: Figure = None):
+    """ change the figure size to the given dimensions. Optionally define if to remove or add space at the top or bottom
+        and left or right.
+    """
     if fig is None:
         fig = plt.gcf()
     oldw, oldh = fig.get_size_inches()
@@ -101,7 +108,9 @@ def changeFigureSize(w, h, cut_from_top=False, cut_from_left=False, fig=None):
                 text.set_position([x0 * fx, 1 - (1 - y0) * fy])
     fig.set_size_inches(w, h, forward=True)
 
-def removeContentFromFigure(fig):
+
+def removeContentFromFigure(fig: Figure):
+    """ remove axes and text from a figure """
     axes = []
     for ax in fig._axstack.as_list():
         axes.append(ax)
@@ -110,7 +119,9 @@ def removeContentFromFigure(fig):
     fig.texts = []
     return axes + text
 
-def addContentToFigure(fig, axes):
+
+def addContentToFigure(fig: Figure, axes: Sequence):
+    """ add axes and texts to a figure """
     index = len(fig._axstack.as_list())
     for ax in axes:
         if isinstance(ax, Axes):
@@ -119,7 +130,9 @@ def addContentToFigure(fig, axes):
         else:
             fig.texts.append(ax)
 
-def imShowFullFigure(im, filename, fig1, dpi):
+
+def imShowFullFigure(im: np.ndarray, filename: str, fig1: Figure, dpi: int):
+    """ create a new axes and display an image in this axes """
     from matplotlib import rcParams
     if dpi is None:
         dpi = rcParams['figure.dpi']
@@ -147,7 +160,11 @@ class changeFolder:
     def __exit__(self, type, value, traceback):
         os.chdir(self.old_dir)
 
-def loadFigureFromFile(filename, fig1=None, offset=None, dpi=None, cache=False):
+
+def loadFigureFromFile(filename: str, fig1: Figure = None, offset: Figure = None, dpi: int = None, cache: bool = False):
+    """ add contents to the current figure from the file defined by filename. It can be either a python script defining
[…]
