+++ pylustrator/exception_swallower.py
@@ -4,6 +4,7 @@ from matplotlib.axis import Axis
 
 
 class Dummy:
+    """ a dummy object that provides dummy attributes, dummy items and dummy returns """
     def __getattr__(self, item):
         return Dummy()
 
@@ -15,6 +16,7 @@ class Dummy:
 
 
 class SaveList(list):
+    """ a list that returns dummy objects when an invalid item is requested """
     def __init__(self, target):
         list.__init__(self, target)
 
@@ -26,6 +28,7 @@ class SaveList(list):
 
 
 class SaveDict(dict):
+    """ a dictionary that returns dummy objects when an invalid item is requested """
     def __init__(self, target):
         dict.__init__(self, target)
 
@@ -37,6 +40,7 @@ class SaveDict(dict):
 
 
 class SaveTuple(tuple):
+    """ a tuple that returns dummy objects when an invalid item is requested """
     def __init__(self, target):
         tuple.__init__(self, target)
 
@@ -48,6 +52,7 @@ class SaveTuple(tuple):
 
 
 class SaveListDescriptor:
+    """ a descriptor that wraps the target value with a SaveList, SaveDict or SaveTuple """
     def __init__(self, variable_name):
         self.variable_name = variable_name
 
@@ -71,16 +76,22 @@ class SaveListDescriptor:
 
 
 def get_axes(self):
+    """ a function that returns the axes of a figure as a SaveList """
     return SaveList(self._axstack.as_list())
 
 
 def return_save_list(func):
+    """ a decorator to wrap the output of a function as a SaveList """
     def wrap(*args, **kwargs):
         return SaveList(func(*args, **kwargs))
     return wrap
 
 
 def swallow_get_exceptions():
+    """ replace lists with lists that return dummy objects when items are not present.
+        this is to ensure that the pylustrator generated code does not fail, even if the user removes some elements
+        from the figure.
+    """
     Figure._get_axes = get_axes
     Figure.axes = property(fget=get_axes)
     Figure.ax_dict = SaveListDescriptor("ax_dict")
