### 2020-02-07 5e169bc
added docstrings to jupyter_cells.py
