### 2020-02-07 bdc09b6
added docstrings to QtShortCuts.py
