+++ pylustrator/QtShortCuts.py
@@ -21,33 +21,31 @@
 
 from qtpy import QtCore, QtGui, QtWidgets
 import numpy as np
+from matplotlib import pyplot as plt
 import matplotlib as mpl
-
-
-def getColorFromCoordinates(p):
-    desktop = QtWidgets.QApplication.desktop()
-    pixmap = QtWidgets.QGuiApplication.screens().at(desktop.screenNumber()).grabWindow(desktop.winId(),
-                                                                                         p.x(), p.y(), 1, 1)
-    i = pixmap.toImage()
-    return i.pixel(0, 0)
 
 
 """ Color Chooser """
 
 class QDragableColor(QtWidgets.QLineEdit):
+    """ a color widget that can be dragged onto another QDragableColor widget to exchange the two colors.
+    alternatively it can be right-clicked to select a color.
+    """
+
     color_changed = QtCore.Signal(str)
 
-    def __init__(self, value):
-        QtWidgets.QLineEdit.__init__(self, value)
+    def __init__(self, value: str):
+        """ initialize with a color """
+        super().__init__(value)
         import matplotlib.pyplot as plt
         self.maps = plt.colormaps()
         self.setAcceptDrops(True)
         self.setAlignment(QtCore.Qt.AlignHCenter)
         self.setColor(value, True)
 
-    def getBackground(self):
-        import matplotlib.pyplot as plt
-        import matplotlib as mpl
+    def getBackground(self) -> str:
+        """ get the background of the color button """
+
         try:
             cmap = plt.get_cmap(self.color)
         except:
@@ -60,7 +58,8 @@ class QDragableColor(QtWidgets.QLineEdit):
         text = text[:-2] + ");"
         return text
 
-    def setColor(self, value, no_signal=False):
+    def setColor(self, value: str, no_signal=False):
+        """ set the current color """
         # display and save the new color
         self.color = value
         self.setText(value)
@@ -70,11 +69,14 @@ class QDragableColor(QtWidgets.QLineEdit):
         else:
             self.setStyleSheet("text-align: center; background-color: %s; border: 2px solid black" % value)
 
-    def getColor(self):
+    def getColor(self) -> str:
+        """ get teh current color """
         # return the color
         return self.color
 
     def mousePressEvent(self, event):
+        """ when a mouse button is pressed """
+        # a left mouse button lets the user drag the color
         if event.button() == QtCore.Qt.LeftButton:
             drag = QtGui.QDrag(self)
             drag.setPixmap(self.grab())
@@ -91,10 +93,12 @@ class QDragableColor(QtWidgets.QLineEdit):
                 self.setStyleSheet("text-align: center; border: 2px solid black; "+self.getBackground())
             else:
                 self.setStyleSheet("text-align: center; background-color: %s; border: 2px solid black" % self.color)
+        # a right mouse button opens a color choose menu
         elif event.button() == QtCore.Qt.RightButton:
             self.openDialog()
 
     def dragEnterEvent(self, event):
+        """ when a color widget is dragged over the current widget """
         if event.mimeData().hasFormat("text/plain") and event.source() != self:
             event.acceptProposedAction()
             if self.color in self.maps:
@@ -103,24 +107,23 @@ class QDragableColor(QtWidgets.QLineEdit):
                 self.setStyleSheet("background-color: %s; border: 2px solid red" % self.color)
 
     def dragLeaveEvent(self, event):
+        """ when the color widget which is dragged leaves the area of this widget """
         if self.color in self.maps:
             self.setStyleSheet("border: 2px solid black; "+self.getBackground())
         else:
             self.setStyleSheet("background-color: %s; border: 2px solid black" % self.color)
 
[…]
