### 2020-02-09 #1989 @labarba
👋 @story645 — It looks like you haven't gotten a start in reviewing this software? We have progress with the second reviewer (Thomas), so I encourage you to come have look now, while the authors may start their revisions.
