### 2020-02-10 fe49ff1
added docstrings to lab_colormap.py and parse_svg.py
