+++ pylustrator/parse_svg.py
@@ -3,6 +3,9 @@ import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
 import matplotlib.transforms as mtransforms
+import matplotlib.path as mpath
+from matplotlib.textpath import TextPath
+from matplotlib.font_manager import FontProperties
 import sys
 import numpy as np
 import re
@@ -11,10 +14,14 @@ import base64
 import matplotlib.text
 from .arc2bez import arcToBezier
 
-def deform(base_trans, x, y, sx=0, sy=0):
+
+def deform(base_trans: mtransforms.Transform, x: float, y: float, sx: float = 0, sy: float = 0):
+    """ apply an affine transformation to the given transformation """
     return mtransforms.Affine2D([[x, sx, 0], [sy, y, 0], [0, 0, 1]]) + base_trans
 
-def parseTransformation(transform_text):
+
+def parseTransformation(transform_text: str) -> mtransforms.Transform:
+    """ convert a transform string in the svg file to a matplotlib transformation """
     base_trans = mtransforms.IdentityTransform()
     if transform_text is None or transform_text == "":
         return base_trans
@@ -53,7 +60,9 @@ def parseTransformation(transform_text):
             print("ERROR: unknown transformation", transform_text)
     return base_trans
 
-def get_inline_style(node, base_style=None):
+
+def get_inline_style(node: minidom.Element, base_style: dict = None) -> dict:
+    """ update the basestyle with the style defined by the style property of the node """
     style = {}
     if base_style is not None:
         style.update(base_style)
@@ -70,7 +79,11 @@ def get_inline_style(node, base_style=None):
         style[key] = value
     return style
 
-def get_css_style(node, css_list, base_style):
+
+def get_css_style(node: minidom.Element, css_list: list, base_style: dict) -> dict:
+    """ update the base_style with the style definitions from the stylesheet that are applicable to the node
+        defined by the classes or id of the node
+    """
     style = {}
     if base_style is not None:
         style.update(base_style)
@@ -85,7 +98,9 @@ def get_css_style(node, css_list, base_style):
             style.update(css_style)
     return style
 
-def apply_style(style, patch):
+
+def apply_style(style: dict, patch: mpatches.Patch) -> dict:
+    """ apply the properties defined in style to the given patch """
     fill_opacity = float(style.get("opacity", 1)) * float(style.get("fill-opacity", 1))
     stroke_opacity = float(style.get("opacity", 1)) * float(style.get("stroke-opacity", 1))
 
@@ -183,8 +198,9 @@ def apply_style(style, patch):
             print("ERROR: could not set style", key, value, file=sys.stderr)
     return style
 
-def font_properties_from_style(style):
-    from matplotlib.font_manager import FontProperties
+
+def font_properties_from_style(style: dict) -> FontProperties:
+    """ convert a style to a FontProperties object """
     fp = FontProperties()
     for key, value in style.items():
         if key == "font-family":
@@ -201,12 +217,16 @@ def font_properties_from_style(style):
             fp.set_stretch(value)
     return fp
 
-def styleNoDisplay(style):
+
+def styleNoDisplay(style: dict) -> bool:
+    """ check whether the style defines not to display the element """
     return style.get("display", "inline") == "none" or \
             style.get("visibility", "visible") == "hidden" or \
             style.get("visibility", "visible") == "collapse"
 
-def plt_patch(node, trans_parent_trans, style, constructor, ids, no_draw=False):
+
[…]
