+++ pylustrator/lab_colormap.py
@@ -21,28 +21,35 @@
 
 """ Colormap """
 import numpy as np
+from typing import Sequence, Union
 from matplotlib.colors import Colormap, ListedColormap, to_rgb
-#from skimage.color import rgb2lab, lab2rgb
 
 
 class CmapColor(list):
+    """ a class that appends metadate to a color """
     def setMeta(self, value, cmap):
         self.value = value
         self.cmap = cmap
 
 
-def convert_rgb2lab(colors):
+def convert_rgb2lab(colors: Sequence):
+    """ convert colors from rgb to lab color space """
+    from skimage.color import rgb2lab
     return [rgb2lab(np.array(c)[None, None, :3]) for c in colors]
 
 
 def convert_lab2rgb(colors):
+    """ convert colors from lab to rgb color space """
+    from skimage.color import lab2rgb
     return [lab2rgb(np.array(c))[0, 0, :3] for c in colors]
 
 
 class LabColormap(ListedColormap):
+    """ a custom colormap that blends between N custom colors """
     init_colors = None
 
-    def __init__(self, colors, N, stops=None):
+    def __init__(self, colors: Sequence, N: int, stops=None):
+        """ initialize with the given colors and stops """
         # store stops
         self.stops = stops
         # set colors
@@ -51,6 +58,7 @@ class LabColormap(ListedColormap):
         Colormap.__init__(self, "test", N)
 
     def _init(self):
+        """ generate the colormap from the given colors (used by ListedColormap) """
         # convert to lab
         lab_colors = convert_rgb2lab(self.init_colors)
         # initialize new list
@@ -66,7 +74,8 @@ class LabColormap(ListedColormap):
         # initialize a listed colormap
         ListedColormap._init(self)
 
-    def __call__(self, value, *args, **kwargs):
+    def __call__(self, value: float, *args, **kwargs):
+        """ get the color associated with the given value from the colormap """
         # get the color
         result = Colormap.__call__(self, value, *args, **kwargs)
         # add meta values to it
@@ -75,11 +84,13 @@ class LabColormap(ListedColormap):
         # return the color
         return result
 
-    def get_color(self):
+    def get_color(self) -> Sequence:
+        """ return all the colors """
         # return the colors
         return self.init_colors
 
-    def set_color(self, color, index=None):
+    def set_color(self, color: Union[str, Sequence], index: int = None):
+        """ set a color to the given index """
         # update the color according to the index
         if index is not None:
             self.init_colors[index] = to_rgb(color)
@@ -92,7 +103,8 @@ class LabColormap(ListedColormap):
         # notify that we have to reinitialize the colormap
         self._isinit = False
 
-    def get_stops(self):
+    def get_stops(self) -> Sequence:
+        """ return the stops """
         # get the stops
         stops = self.stops
         # if they are not defined, interpolate from 0 to 1
@@ -102,6 +114,7 @@ class LabColormap(ListedColormap):
         return stops
 
     def linearize_lightness(self):
+        """ linearize the lightness of the colors in the colormap """
         # convert to lab
         lab_colors = convert_rgb2lab(self.init_colors)
         # define start and end lightness
