+++ docs/conf.py
@@ -26,8 +26,22 @@ import shlex
 print(os.getcwd())
 
 import mock
-MOCK_MODULES = ['imageio', 'peewee', 'PIL', 'PIL._util', 'numpy', 'qimage2ndarray', 'sip', 'PyQt4', 'PyQt4.QtGui', 'qtpy', 'qtpy.QtGui', 'qtpy.QtCore', 'qtpy.QtWidgets', 'qtawesome', 'matplotlib', 'matplotlib.backends', 'matplotlib.backends.backend_qt4agg', 'matplotlib.backends.backend_qt4', 'matplotlib.figure']
-sys.modules.update((mod_name, mock.MagicMock()) for mod_name in MOCK_MODULES)
+# try to import the modules of the package and mock everything that is not found
+while True:
+    try:
+        # here are the modules that should be imported for the documentation
+        import pylustrator
+    # if an import error occurs
+    except ImportError as err:
+        # get the module name from the error message
+        name = str(err).split("'")[1]
+        print("Mock:", name)
+        # and mock it
+        sys.modules.update((mod_name, mock.MagicMock()) for mod_name in [name])
+        # then try again to import it
+        continue
+    else:
+        break
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
