+++ docs/index.rst
@@ -111,3 +111,4 @@ account.
 
    styling
    composing
+   api
