+++ docs/composing.rst
@@ -1,3 +1,4 @@
+.. _composing:
 
 Composing Figures
 =================
