+++ docs/api.rst
@@ -1,0 +1,9 @@
+API
+===
+
+The API of pylustrator is kept quite simple. Most interaction with the pylustrator package is by using the interactive
+interface.
+
+.. autofunction:: pylustrator.start
+
+.. autofunction:: pylustrator.load
