+++ docs/conf.py
@@ -25,6 +25,11 @@ import shlex
 
 print(os.getcwd())
 
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+
 import mock
 # try to import the modules of the package and mock everything that is not found
 while True:
@@ -42,11 +47,6 @@ while True:
         continue
     else:
         break
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath(os.path.join("..")))
 
 # -- General configuration ------------------------------------------------
 
