### 2020-02-10 25d88b2
fixed making matplotlibwidget mockable for docs
