### 2020-02-10 99fceb0
added matplotlib to the requirements.readthedocs.txt
