+++ docs/requirements.readthedocs.txt
@@ -1,2 +1,3 @@
 sphinxcontrib-bibtex
 nbsphinx
+matplotlib
