### 2020-02-10 95e86da
added pyside2, qtpy to the requirements.readthedocs.txt
