+++ docs/requirements.readthedocs.txt
@@ -1,3 +1,5 @@
 sphinxcontrib-bibtex
 nbsphinx
 matplotlib
+pyside2
+qtpy
