### 2020-02-10 e430d4b
fixed error in .readthedocs.yml
