+++ .readthedocs.yml
@@ -1,4 +1,3 @@
- .readthedocs.yml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 
