### 2020-02-10 042a63d
fixed error in .readthedocs.yml
