+++ .readthedocs.yml
@@ -15,4 +15,4 @@ formats: all
 python:
   version: 3.7
   install:
-    - requirements: docs/requirements.txt
+    - requirements: docs/requirements.readthedocs.txt
