### 2020-02-12 4862403
Moved parts of QtGuiDrag.py to other files
