+++ pylustrator/QLinkableWidgets.py
@@ -1,0 +1,609 @@
+import os
+import sys
+import traceback
+
+import qtawesome as qta
+from matplotlib import _pylab_helpers
+
+from .QComplexWidgets import *
+from .ax_rasterisation import rasterizeAxes, restoreAxes
+from .change_tracker import setFigureVariableNames
+from .drag_helper import DragManager
+from .exception_swallower import swallow_get_exceptions
+from .matplotlibwidget import MatplotlibWidget
+
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+import matplotlib.transforms as transforms
+import numpy as np
+from matplotlib.artist import Artist
+from matplotlib.figure import Figure
+from qtpy import QtCore, QtWidgets, QtGui
+
+
+class Linkable:
+    """ a class that automatically links a widget with the property of a matplotlib artist
+    """
+
+    def link(self, property_name: str, signal: QtCore.Signal = None, condition: callable = None, direct: bool = False):
+        self.element = None
+        self.direct = direct
+        if direct:
+            parts = property_name.split(".")
+            s = self
+            def get():
+                target = s.element
+                for part in parts:
+                    target = getattr(target, part)
+                return target
+            def set(v):
+                get()
+                target = s.element
+                for part in parts[:-1]:
+                    target = getattr(target, part)
+                setattr(target, parts[-1], v)
+                return [s.element]
+            self.setLinkedProperty = set
+            self.getLinkedProperty = get
+            self.serializeLinkedProperty = lambda x: "." + property_name + " = %s" % x
+        else:
+            def set(v):
+                elements = []
+                getattr(self.element, "set_" + property_name)(v)
+                elements.append(self.element)
+                for elm in self.element.figure.selection.targets:
+                    elm = elm.target
+                    if elm != self.element:
+                        try:
+                            getattr(elm, "set_"+property_name, None)(v)
+                        except TypeError as err:
+                            pass
+                        else:
+                            elements.append(elm)
+                return elements
+            self.setLinkedProperty = set#lambda text: getattr(self.element, "set_"+property_name)(text)
+            self.getLinkedProperty = lambda: getattr(self.element, "get_"+property_name)()
+            self.serializeLinkedProperty = lambda x: ".set_"+property_name+"(%s)" % x
+
+        if condition is None:
+            self.condition = lambda x: True
+        else:
+            self.condition = condition
+
+        self.editingFinished.connect(self.updateLink)
+        signal.connect(self.setTarget)
+
+    def setTarget(self, element: Artist):
+        """ set the target for the widget """
+        self.element = element
+        try:
+            self.set(self.getLinkedProperty())
+            self.setEnabled(self.condition(element))
+        except AttributeError:
+            self.hide()
+        else:
+            self.show()
+
+    def updateLink(self):
+        """ update the linked property """
+        try:
+            elements = self.setLinkedProperty(self.get())
+        except AttributeError:
+            return
+        for element in elements:
+            if isinstance(element, mpl.figure.Figure):
+                fig = element
+            else:
+                fig = element.figure
+            fig.change_tracker.addChange(element, self.serializeLinkedProperty(self.getSerialized()))
+        fig.canvas.draw()
+
+    def set(self, value):
+        """ set the value (to be overloaded) """
+        pass
+
[…]
