+++ pylustrator/QComplexWidgets.py
@@ -1,0 +1,1007 @@
+import os
+import sys
+
+import qtawesome as qta
+from matplotlib import _pylab_helpers
+from matplotlib.axes._subplots import Axes
+
+
+from .QLinkableWidgets import *
+
+
+class TextPropertiesWidget(QtWidgets.QWidget):
+    stateChanged = QtCore.Signal(int, str)
+    noSignal = False
+    target_list = None
+
+    def __init__(self, layout):
+        QtWidgets.QWidget.__init__(self)
+        layout.addWidget(self)
+        self.layout = QtWidgets.QHBoxLayout(self)
+        self.layout.setContentsMargins(0, 0, 0, 0)
+
+        self.buttons_align = []
+        self.align_names = ["left", "center", "right"]
+        for align in self.align_names:
+            button = QtWidgets.QPushButton(qta.icon("fa.align-" + align), "")
+            button.setCheckable(True)
+            button.clicked.connect(lambda x, name=align: self.changeAlign(name))
+            self.layout.addWidget(button)
+            self.buttons_align.append(button)
+
+        self.button_bold = QtWidgets.QPushButton(qta.icon("fa.bold"), "")
+        self.button_bold.setCheckable(True)
+        self.button_bold.clicked.connect(self.changeWeight)
+        self.layout.addWidget(self.button_bold)
+
+        self.button_italic = QtWidgets.QPushButton(qta.icon("fa.italic"), "")
+        self.button_italic.setCheckable(True)
+        self.button_italic.clicked.connect(self.changeStyle)
+        self.layout.addWidget(self.button_italic)
+
+        self.button_color = QColorWidget(self.layout)
+        self.button_color.valueChanged.connect(self.changeColor)
+
+        self.layout.addStretch()
+
+        self.font_size = QtWidgets.QSpinBox()
+        self.layout.addWidget(self.font_size)
+        self.font_size.valueChanged.connect(self.changeFontSize)
+
+        self.label = QtWidgets.QPushButton(qta.icon("fa.font"), "")  # .pixmap(16))
+        self.layout.addWidget(self.label)
+        self.label.clicked.connect(self.selectFont)
+
+        self.button_delete = QtWidgets.QPushButton(qta.icon("fa.trash"), "")
+        self.button_delete.clicked.connect(self.delete)
+        self.layout.addWidget(self.button_delete)
+
+    def convertMplWeightToQtWeight(self, weight):
+        weight_dict = {'normal': QtGui.QFont.Normal, 'bold': QtGui.QFont.Bold, 'heavy': QtGui.QFont.ExtraBold,
+                       'light': QtGui.QFont.Light, 'ultrabold': QtGui.QFont.Black, 'ultralight': QtGui.QFont.ExtraLight}
+        if weight in weight_dict:
+            return weight_dict[weight]
+        return weight_dict["normal"]
+
+    def convertQtWeightToMplWeight(self, weight):
+        weight_dict = {QtGui.QFont.Normal: 'normal', QtGui.QFont.Bold: 'bold', QtGui.QFont.ExtraBold: 'heavy',
+                       QtGui.QFont.Light: 'light', QtGui.QFont.Black: 'ultrabold', QtGui.QFont.ExtraLight: 'ultralight'}
+        if weight in weight_dict:
+            return weight_dict[weight]
+        return "normal"
+
+    def selectFont(self):
+        font0 = QtGui.QFont()
+        font0.setFamily(self.target.get_fontname())
+        font0.setWeight(self.convertMplWeightToQtWeight(self.target.get_weight()))
+        font0.setItalic(self.target.get_style() == "italic")
+        font0.setPointSizeF(self.target.get_fontsize())
+        font, x = QtWidgets.QFontDialog.getFont(font0, self)
+
[…]

[…]
