### 2020-02-13 6135f0d
QLinkableWidgets added docstrings
