+++ pylustrator/QLinkableWidgets.py
@@ -1,16 +1,4 @@
-import os
-import sys
-import traceback
-
-import qtawesome as qta
-from matplotlib import _pylab_helpers
-
-from .QComplexWidgets import *
-from .ax_rasterisation import rasterizeAxes, restoreAxes
-from .change_tracker import setFigureVariableNames
-from .drag_helper import DragManager
-from .exception_swallower import swallow_get_exceptions
-from .matplotlibwidget import MatplotlibWidget
+from typing import Optional, Sequence
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
@@ -31,11 +19,13 @@ class Linkable:
         if direct:
             parts = property_name.split(".")
             s = self
+
             def get():
                 target = s.element
                 for part in parts:
                     target = getattr(target, part)
                 return target
+
             def set(v):
                 get()
                 target = s.element
@@ -43,6 +33,7 @@ class Linkable:
                     target = getattr(target, part)
                 setattr(target, parts[-1], v)
                 return [s.element]
+
             self.setLinkedProperty = set
             self.getLinkedProperty = get
             self.serializeLinkedProperty = lambda x: "." + property_name + " = %s" % x
@@ -55,15 +46,16 @@ class Linkable:
                     elm = elm.target
                     if elm != self.element:
                         try:
-                            getattr(elm, "set_"+property_name, None)(v)
+                            getattr(elm, "set_" + property_name, None)(v)
                         except TypeError as err:
                             pass
                         else:
                             elements.append(elm)
                 return elements
-            self.setLinkedProperty = set#lambda text: getattr(self.element, "set_"+property_name)(text)
-            self.getLinkedProperty = lambda: getattr(self.element, "get_"+property_name)()
-            self.serializeLinkedProperty = lambda x: ".set_"+property_name+"(%s)" % x
+
+            self.setLinkedProperty = set  # lambda text: getattr(self.element, "set_"+property_name)(text)
+            self.getLinkedProperty = lambda: getattr(self.element, "get_" + property_name)()
+            self.serializeLinkedProperty = lambda x: ".set_" + property_name + "(%s)" % x
 
         if condition is None:
             self.condition = lambda x: True
@@ -112,17 +104,23 @@ class Linkable:
 
 
 class FreeNumberInput(QtWidgets.QLineEdit):
-    """ like a QSpinBox for number import, but without min or max range or a fixed resolution.
-    Especially important for the limits of logarithmic plots.
-    """
     send_signal = True
     valueChanged = QtCore.Signal(float)
 
     def __init__(self):
+        """ Like a QSpinBox for number import, but without min or max range or a fixed resolution.
+        Especially important for the limits of logarithmic plots.
+
+        Attributes:
+            send_signal : Whether to currently emit the valueChanged signal or not (to prevent the signal from being
+                emitted when the value is set by script.
+            valueChanged : a signal that is emitted when the value is changed by the user
+        """
         QtWidgets.QLineEdit.__init__(self)
         self.textChanged.connect(self.emitValueChanged)
 
-    def emitValueChanged(self, value):
+    def emitValueChanged(self):
+        """ connected to the textChanged signal """
         if self.send_signal:
             try:
                 value = self.value()
@@ -132,7 +130,8 @@ class FreeNumberInput(QtWidgets.QLineEdit):
                 self.setStyleSheet("background: #d56060; border: red")
                 pass
 
-    def value(self):
+    def value(self) -> Optional[float]:
+        """ return the value of the input field """
         try:
             return float(self.text())
         except ValueError:
[…]
