### 2020-02-14 6bb7166
QComplexWidgets.py added docstrings
