+++ pylustrator/QComplexWidgets.py
@@ -24,7 +24,10 @@ import numpy as np
 from matplotlib.artist import Artist
 from matplotlib.figure import Figure
 from qtpy import QtCore, QtWidgets, QtGui
+from typing import Any, Optional
 
+from .change_tracker import getReference
+from .helper_functions import changeFigureSize
 from .QLinkableWidgets import QColorWidget, CheckWidget, TextWidget, RadioWidget, DimensionsWidget, NumberWidget, ComboWidget
 
 
@@ -33,7 +36,12 @@ class TextPropertiesWidget(QtWidgets.QWidget):
     noSignal = False
     target_list = None
 
-    def __init__(self, layout):
+    def __init__(self, layout: QtWidgets.QLayout):
+        """ A widget to edit the properties of a Matplotlib text
+
+        Args:
+            layout: the layout to which to add the widget
+        """
         QtWidgets.QWidget.__init__(self)
         layout.addWidget(self)
         self.layout = QtWidgets.QHBoxLayout(self)
@@ -75,14 +83,16 @@ class TextPropertiesWidget(QtWidgets.QWidget):
         self.button_delete.clicked.connect(self.delete)
         self.layout.addWidget(self.button_delete)
 
-    def convertMplWeightToQtWeight(self, weight):
+    def convertMplWeightToQtWeight(self, weight: str) -> int:
+        """ convert a font weight string to a weight enumeration of Qt """
         weight_dict = {'normal': QtGui.QFont.Normal, 'bold': QtGui.QFont.Bold, 'heavy': QtGui.QFont.ExtraBold,
                        'light': QtGui.QFont.Light, 'ultrabold': QtGui.QFont.Black, 'ultralight': QtGui.QFont.ExtraLight}
         if weight in weight_dict:
             return weight_dict[weight]
         return weight_dict["normal"]
 
-    def convertQtWeightToMplWeight(self, weight):
+    def convertQtWeightToMplWeight(self, weight: int) -> str:
+        """ convert a Qt weight value to a string for use in matmplotlib """
         weight_dict = {QtGui.QFont.Normal: 'normal', QtGui.QFont.Bold: 'bold', QtGui.QFont.ExtraBold: 'heavy',
                        QtGui.QFont.Light: 'light', QtGui.QFont.Black: 'ultrabold', QtGui.QFont.ExtraLight: 'ultralight'}
         if weight in weight_dict:
@@ -90,6 +100,7 @@ class TextPropertiesWidget(QtWidgets.QWidget):
         return "normal"
 
     def selectFont(self):
+        """ open a font select dialog """
         font0 = QtGui.QFont()
         font0.setFamily(self.target.get_fontname())
         font0.setWeight(self.convertMplWeightToQtWeight(self.target.get_weight()))
@@ -118,7 +129,8 @@ class TextPropertiesWidget(QtWidgets.QWidget):
         self.target.figure.canvas.draw()
         self.setTarget(self.target_list)
 
-    def setTarget(self, element):
+    def setTarget(self, element: Artist):
+        """ set the target artist for this widget """
         if isinstance(element, list):
             self.target_list = element
             element = element[0]
@@ -141,6 +153,7 @@ class TextPropertiesWidget(QtWidgets.QWidget):
         self.target = element
 
     def delete(self):
+        """ delete the target text """
         if self.target is not None:
             fig = self.target.figure
             fig.change_tracker.removeElement(self.target)
@@ -148,7 +161,8 @@ class TextPropertiesWidget(QtWidgets.QWidget):
             # self.target.set_visible(False)
             fig.canvas.draw()
 
-    def changeWeight(self, checked):
+    def changeWeight(self, checked: bool):
+        """ set bold or normal """
         if self.target:
[…]
