### 2020-02-14 a9408d0
QtGuiDrag.py added docstrings
