+++ pylustrator/QtGuiDrag.py
@@ -197,19 +197,15 @@ def warnAboutTicks(fig):
 
 
 class myTreeWidgetItem(QtGui.QStandardItem):
-    def __init__(self, parent=None):
+    def __init__(self, parent: QtWidgets.QWidget = None):
+        """ a tree view item to display the contents of the figure """
         QtGui.QStandardItem.__init__(self, parent)
 
-    def __lt__(self, otherItem):
+    def __lt__(self, otherItem: QtGui.QStandardItem):
+        """ how to sort the items """
         if self.sort is None:
             return 0
         return self.sort < otherItem.sort
-        column = self.treeWidget().sortColumn()
-
-        if column == 0 or column == 6 or column == 7 or column == 8:
-            return float(self.text(column)) < float(otherItem.text(column))
-        else:
-            return self.text(column) < otherItem.text(column)
 
 
 class MyTreeView(QtWidgets.QTreeView):
@@ -222,7 +218,14 @@ class MyTreeView(QtWidgets.QTreeView):
     last_selection = None
     last_hover = None
 
-    def __init__(self, parent, layout, fig):
+    def __init__(self, parent: QtWidgets.QWidget, layout: QtWidgets.QLayout, fig: Figure):
+        """ A tree view to display the contents of a figure
+
+        Args:
+            parent: the parent widget
+            layout: the layout to which to add the tree view
+            fig: the target figure
+        """
         super().__init__()
 
         self.fig = fig
@@ -257,7 +260,8 @@ class MyTreeView(QtWidgets.QTreeView):
 
         self.expand(None)
 
-    def selectionChanged(self, selection, y):
+    def selectionChanged(self, selection: QtCore.QItemSelection, y: QtCore.QItemSelection):
+        """ when the selection in the tree view changes """
         try:
             entry = selection.indexes()[0].model().itemFromIndex(selection.indexes()[0]).entry
         except IndexError:
@@ -266,7 +270,8 @@ class MyTreeView(QtWidgets.QTreeView):
             self.last_selection = entry
             self.item_selected(entry)
 
-    def setCurrentIndex(self, entry):
+    def setCurrentIndex(self, entry: Artist):
+        """ set the currently selected entry """
         while entry:
             item = self.getItemFromEntry(entry)
             if item is not None:
@@ -277,17 +282,17 @@ class MyTreeView(QtWidgets.QTreeView):
             except AttributeError:
                 return
 
-    def treeClicked(self, index):
-        # upon selecting one of the tree elements
+    def treeClicked(self, index: QtCore.QModelIndex):
+        """ upon selecting one of the tree elements """
         data = index.model().itemFromIndex(index).entry
         return self.item_clicked(data)
 
-    def treeActivated(self, index):
-        # upon selecting one of the tree elements
+    def treeActivated(self, index: QtCore.QModelIndex):
+        """ upon selecting one of the tree elements """
         data = index.model().itemFromIndex(index).entry
         return self.item_activated(data)
 
-    def eventFilter(self, object, event):
+    def eventFilter(self, object: QtWidgets.QWidget, event: QtCore.QEvent):
         """ event filter for tree view port to handle mouse over events and marker highlighting"""
         if event.type() == QtCore.QEvent.HoverMove:
             index = self.indexAt(event.pos())
@@ -314,33 +319,39 @@ class MyTreeView(QtWidgets.QTreeView):
 
         return False
 
-    def queryToExpandEntry(self, entry):
+    def queryToExpandEntry(self, entry: Artist) -> list:
+        """ when expanding a tree item """
         if entry is None:
             return [self.fig]
         return entry.get_children()
 
-    def getParentEntry(self, entry):
[…]
