### 2020-02-14 5f06cbb
QtGui.py added docstrings
