+++ pylustrator/QtGui.py
@@ -19,15 +19,22 @@
 # You should have received a copy of the GNU General Public License
 # along with Pylustrator. If not, see <http://www.gnu.org/licenses/>
 
-from __future__ import division, print_function
 from qtpy import QtCore, QtWidgets, QtGui
 import qtawesome as qta
 
 import numpy as np
 import matplotlib.pyplot as plt
-from matplotlib.backends.backend_qt4 import NavigationToolbar2QT as NavigationToolbar
+from qtpy import API_NAME as QT_API_NAME
+if QT_API_NAME.startswith("PyQt4"):
+    from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as Canvas
+    from matplotlib.backends.backend_qt4 import NavigationToolbar2QT as NavigationToolbar
+else:
+    from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as Canvas
+    from matplotlib.backends.backend_qt5 import NavigationToolbar2QT as NavigationToolbar
 from .matplotlibwidget import MatplotlibWidget
 from matplotlib import _pylab_helpers
+from matplotlib.figure import Figure
+from matplotlib.artist import Artist
 import matplotlib as mpl
 import qtawesome as qta
 
@@ -48,6 +55,7 @@ app = None
 
 
 def initialize():
+    """ patch figure and show to display the color chooser GUI """
     global app
     if app is None:
         app = QtWidgets.QApplication(sys.argv)
@@ -56,6 +64,7 @@ def initialize():
 
 
 def show():
+    """ the patched show to display the color choose gui """
     global figures
     # iterate over figures
     for figure in figures:
@@ -68,6 +77,7 @@ def show():
 
 
 def figure(num=None, figsize=None, *args, **kwargs):
+    """ the patched figure to initialize to color chooser GUI """
     global figures
     # if num is not defined create a new number
     if num is None:
@@ -87,10 +97,12 @@ def figure(num=None, figsize=None, *args, **kwargs):
     # return the figure
     return canvas.figure
 
+
 """ Figure list functions """
 
 
-def addChildren(color_artists, parent):
+def addChildren(color_artists: list, parent: Artist):
+    """ find all the children of an Artist that use a color """
     for artist in parent.get_children():
         # ignore empty texts
         if isinstance(artist, mpl.text.Text) and artist.get_text() == "":
@@ -157,11 +169,15 @@ def addChildren(color_artists, parent):
                     # add the artist
                     color_artists[color].append([color_type_name, artist, None, None, None])
 
-def figureListColors(figure):
+
+def figureListColors(figure: Figure):
+    """ add all artist with colors to a list in the figure """
     figure.color_artists = {}
     addChildren(figure.color_artists, figure)
 
-def figureSwapColor(figure, new_color, color_base):
+
+def figureSwapColor(figure: Figure, new_color: str, color_base: str):
+    """ swap two colors of a figure """
     if getattr(figure, "color_artists", None) is None:
         figureListColors(figure)
     changed_cmaps = []
@@ -186,24 +202,34 @@ def figureSwapColor(figure, new_color, color_base):
                     continue
             # use the attributes setter method
             getattr(artist, "set_" + color_type_name)(cmap(value))
-            artist.figure.change_tracker.addChange(artist, ".set_" + color_type_name + "(plt.get_cmap(\"%s\")(%s))" % (cmap.name, str(value)))
+            artist.figure.change_tracker.addChange(artist, ".set_" + color_type_name + "(plt.get_cmap(\"%s\")(%s))" % (
+            cmap.name, str(value)))
         else:
             if new_color in maps:
                 cmap = plt.get_cmap(new_color)
                 getattr(artist, "set_" + color_type_name)(cmap(0))
                 artist.figure.change_tracker.addChange(artist,
[…]
