### 2020-02-17 a007af5
added a toolbar similar to the mpl toolbar
