+++ pylustrator/QtGuiDrag.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # QtGuiDrag.py
 
-# Copyright (c) 2016-2019, Richard Gerum
+# Copyright (c) 2016-2020, Richard Gerum
 #
 # This file is part of Pylustrator.
 #
