+++ docs/conf.py
@@ -88,9 +88,9 @@ author = u'Richard Gerum'
 # built documents.
 #
 # The short X.Y version.
-version = '0.13.0'
+version = '0.14.0'
 # The full version, including alpha/beta/rc tags.
-release = '0.13.0'
+release = '0.14.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
