### 2020-05-15 060ceeb
paper: cite paraview, avizo, and draganfly and removed strong novelty claim
