+++ docs/paper.bib
@@ -49,6 +49,31 @@ year = {2016}
   note = {Accessed: 2019-09-19}
 }
 
+@misc{dragonfly,
+  title = {Dragonfly},
+  howpublished = {\url{https://www.theobjects.com/dragonfly/}},
+  note = {Accessed: 2020-05-15}
+}
+
+@inproceedings{westenberger2008avizo,
+  title={AVIZO-3D visualization framework},
+  author={Westenberger, Peter},
+  booktitle={Geoinformatics Conference},
+  pages={1--11},
+  year={2008}
+}
+
+
+@article{ahrens2005paraview,
+  title={Paraview: An end-user tool for large data visualization},
+  author={Ahrens, James and Geveci, Berk and Law, Charles},
+  journal={The visualization handbook},
+  volume={717},
+  year={2005},
+  publisher={Elsevier M{\"u}nchen}
+}
+
+
 @Manual{R,
   title        = {R: A Language and Environment for Statistical
                   Computing},
