+++ docs/paper.md
@@ -27,11 +27,12 @@ In recent years, more and more researchers have called attention to a growing "r
 One major step to overcome these shortcomings is the publication of all raw data as well as a documented version of the code used for evaluation [@baker2016scientists]. The ideal case would be that anyone interested can download the raw data and exactly reproduce the figures of the publication.
 
 To address the issue of data availability, researchers are encouraged to provide their data in online repositories, e.g. dryad [@dryad]. However, this data is useless unless the complete evaluation procedure in the terms of all evaluation and visualisation steps
-can be comprehended by other scientists. The best way to to so is to provide a complete well documented evaluation code, 
+can be comprehended by other scientists. The best way to do so is to provide a complete well documented evaluation code, 
 including all important steps from basic artifact corrections up to the final plot to be published. Open Source scripting languages like Python [@Rossum1995] or R [@R] are ideal for such code as open source languages are accessible for everyone. Furthermore, interpreted languages do not need to be compiled, therefore have less obstacles for the user to run the code. The last part of the evaluation of the data is the visualisation, which is crucial to communicate the results [@Tufte1893]. 
 This paper deals with the visualization step which consists of two parts: the generation of simple plots from data and composing meaningful figures from these plots.
 
-The first part of generating the building blocks of figures, the plots, is already covered in various toolkits, e.g. Matplotlib [@Hunter2007], Bokeh [@Bokeh] or Seaborn [@seaborn]. But to generate reproducible figures from simple plots, no convenient toolkit is yet available. Matplotlib already offers figures composed of multiple subplots, but to generate a complete figure ready for publication a lot of code is needed to add all formatting, annotations and styling commands. Therefore, this approach is often not followed as it is impractical for real applications. Users often prefer graphical tools such as image manipulation software, e.g. GIMP [@GIMP] or Inkscape [@Inkscape]. These offer great flexibility, but cannot provide a reproducible way of generating figures and bear the danger of accidentally changing data points. Also important to note is that by using an image manipulation software, any small change to the evaluation requires to re-edit the figure in the image manipulation software. A process that slows down the creation of figures and is prone to errors.
+The first part of generating the building blocks of figures, the plots, is already covered in various toolkits, e.g. Matplotlib [@Hunter2007], Bokeh [@Bokeh] or Seaborn [@seaborn]. There, already exist some software that records user interactions to generate
+3D plots ([@ahrens2005paraview], [@westenberger2008avizo], [@dragonfly]), but to generate reproducible figures from simple plots scripts, no convenient toolkit is yet available for Python. Matplotlib already offers figures composed of multiple subplots, but to generate a complete figure ready for publication a lot of code is needed to add all formatting, annotations and styling commands. Therefore, this approach is often not followed as it is impractical for real applications. Users often prefer graphical tools such as image manipulation software, e.g. GIMP [@GIMP] or Inkscape [@Inkscape]. These offer great flexibility, but cannot provide a reproducible way of generating figures and bear the danger of accidentally changing data points. Also important to note is that by using an image manipulation software, any small change to the evaluation requires to re-edit the figure in the image manipulation software. A process that slows down the creation of figures and is prone to errors.
 
 # Algorithm and Exemplary Results
 
@@ -66,7 +67,7 @@ It is important to note that the automatically generated code only relies on Mat
 The documentation of ``pylustrator`` can be found on https://pylustrator.readthedocs.org.
 
 # Conclusion
-This study introduces a novel method to create publishable figures from single plots based on an open source Python library called ``pylustrator``. The figures can be arranged by drag and drop and the ``pylustrator`` library produces the according code. Thus, this library provides a valuable contribution to tackle the reproducibility crisis.
[…]
