### 2020-05-15 #1989 @rgerum
Dear Editors and Reviewers,
As already quite some time passed since the first review arrived, I now take the liberty to already respond to the first reviewer before the second review is there.

@tacaswell, Thank you very much for the positivity assessment of the package.

>   Created rgerum/pylustrator#17 to track missing community guide lines
- I added a guilde line paragraph to the readme.md file

>  This software is primarily a GUI interface to editing Matplotlib figures (which is pretty awesome!) and the intended usage is through the UI, not as a library. There is not documentation on the API (because there is not one) nor automated tests.
- I added a section to the documentation to document the functions of the package.

> Poking through the code there is a lack of docstrings on a majority of the functions.
- I added small docstrings to the functions and added type hints to improve readability of the code

> While I don't think that it is worth doing a full GUI mock test, I think it would be reasonable to test some of the functionality (such as ensuring the the monkey patching works as expected, that the file patching works as expected for a manually generated change set etc).
- I added an automatic test that test the basic functions of the package: monkey patching and code generation.

> This sort of user-action-recording functionality exist in several other visualization / image analysis packages (paraview/vtk/and friends, avizo, dragonfly, etc). I know of no other tool that provides this functionality on top of Matplotlib, however I think that the novelty claim in the conclusions of the paper are too strong.
- I removed the strong novenlty claim from the paper and referenced paraview, avizo and dragonfly.

Thank you for your comments that helped to improve the software.
