### 2020-06-26 #1989 @story645
Hi, incredibly sorry my review took so long. Mostly fine, some slight concerns:
1. the documentation homepage is missing a list of major dependencies - I'm not sure if this fits w/ joss guidelines or not
2. the paper needs copy editing (typos, some very clunky constructions) - not sure if this happens before or after the "approve for publication" step
3.  the paragraph starting with `pylustrator tracks all these changes to translate` would be clearer if either the language in it directly matched the figure (for example the figure also had an element labeled `command object`, `target object` etc) or if there was a set of smaller figure walking through these steps.
