### 2020-06-29 7ebb281
docs: added list of dependencies
