### 2020-06-29 d7cc5d7
paper: improved text and rewrote the paragraph of change tracking
