### 2020-06-29 #1989 @rgerum
Thank you for the review.

1. I added a list of dependencies to the documentation page. https://pylustrator.readthedocs.io/en/latest/#installation

2. I did some copy editing of the text.

3. I rewrote this paragraph and I hope it should be more clear now.

You can see the changes to the text here: https://github.com/rgerum/pylustrator/commit/d7cc5d798212c79a54901ad19a71a39a8efbac76
