### 2020-07-13 #1989 @story645
thanks @rgerum for your incredible patience, I think the paper is much clearer now. Two more suggestion and then I think it's ready for publication
missing word: "from this dependency a  object? is generated"
& I think
"when a file with automatically generated code is loaded, pylustrator splits all the ..." might be a drop clearer if you refer back to figure 1 here to explain what this means.
