### 2020-07-20 #1989 @labarba
@story645 — we're missing the check on "quality of writing." did you find issues there?
