+++ docs/conf.py
@@ -88,9 +88,9 @@ author = u'Richard Gerum'
 # built documents.
 #
 # The short X.Y version.
-version = '-v'
+version = '1.0.0'
 # The full version, including alpha/beta/rc tags.
-release = '-v'
+release = '1.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
