### 2020-07-20 #1989 @labarba
I'm reading the paper now, and have minor edits to suggest, which I could enter easily as a PR on the markdown. Before I proceed: I recommend swapping the term "evaluation" for "analysis" in the Background section.
