+++ docs/paper.md
@@ -23,17 +23,16 @@ bibliography: paper.bib
 
 # Background
 
-In recent years, more and more researchers have called attention to a growing "reproducibility crisis" [@CRL16846]. An important factor contributing to problems in the reproducibility of results from published studies is the unavailability of the raw data from the original experiment and the unavailability of the methods or code used for the evaluation the raw data [@Baker2016]. One major step to overcome these shortcomings is the publication of all raw data and a documented version of the code used for evaluation [@baker2016scientists]. The ideal case would be that anyone interested can download the raw data and reproduce the figures of the publication exactly.
+In recent years, more and more researchers have called attention to a growing "reproducibility crisis" [@CRL16846]. An important factor contributing to problems in the reproducibility of results from published studies is the unavailability of the raw data from the original experiment and the unavailability of the methods or code used for the analysis the raw data [@Baker2016]. One major step to overcome these shortcomings is the publication of all raw data and a documented version of the code used for analysis [@baker2016scientists]. The ideal case would be that anyone interested can download the raw data and reproduce the figures of the publication exactly.
 
-To address the issue of data availability, researchers are encouraged to make their data available in online repositories such as dryad [@dryad]. However, these data are useless unless the complete evaluation procedure, in the terms of all evaluation and visualisation steps, can be comprehended by other scientists. The best way to achieve this is to provide a complete, well documented evaluation code, including all important steps from the basic artifact corrections to the final plot to be published.Open source scripting languages such as Python [@Rossum1995] or R [@R] are ideal for such code because open source languages are accessible to everyone. In addition, interpreted languages do not need to be compiled, therefore have less obstacles for the user to run the code. The final part of the data evaluation is the visualisation, which is crucial for communicating the results [@Tufte1893]. This paper deals with the visualization step, which consists of two parts: the generation of simple plots from data and composing meaningful figures from these plots.
+To address the issue of data availability, researchers are encouraged to make their data available in online repositories such as dryad [@dryad]. However, these data are useless unless the complete analysis procedure, in the terms of all analysis and visualisation steps, can be comprehended by other scientists. The best way to achieve this is to provide a complete, well documented analysis code, including all important steps from the basic artifact corrections to the final plot to be published. Open source scripting languages such as Python [@Rossum1995] or R [@R] are ideal for such code because open source languages are accessible to everyone. In addition, interpreted languages do not need to be compiled, therefore have less obstacles for the user to run the code. The final part of the data analysis is the visualisation, which is crucial for communicating the results [@Tufte1893]. This paper deals with the visualization step, which consists of two parts: the generation of simple plots from data and composing meaningful figures from these plots.
 
-The first part of generating the building blocks of figures, the plots, is already covered in various toolkits, e.g. Matplotlib [@Hunter2007], Bokeh [@Bokeh] or Seaborn [@seaborn]. There, already exist some software that records user interactions to generate
[…]
