### 2020-07-20 #1989 @rgerum
I have added the pull request. Luckily Github has a nice diff display for markdown now! I found an additional typo which I corrected. Thank you for improving the paper text.
