### 2020-07-20 #1989 @labarba
@story645 — I've checked the last item for you, as I edited the writing myself now.

@rgerum — please now make an archive of your software repository on Zenodo, edit the Zenodo metadata so the title and author list matches the JOSS paper, and report the archive DOI here.
